pyproject.toml
[tool.poetry]
name = "olas-operate-middleware"
version = "0.10.7"
description = ""
authors = ["David Vilela <dvilelaf@gmail.com>", "Viraj Patel <vptl185@gmail.com>"]
readme = "README.md"
packages = [
    { include = "operate" }
]
include = [
    "operate/data/acn/*",
    "operate/data/contracts/*",
    "operate/data/contracts/staking_token/*",
    "operate/data/contracts/staking_token/build/*",
]

[tool.poetry.scripts]
operate = "operate.cli:main"

[tool.poetry.dependencies]
python = "<3.12,>=3.9"
open-autonomy = "^0.20.2"
open-aea-ledger-cosmos = "==1.65.0"
open-aea-ledger-ethereum = "==1.65.0"
open-aea-ledger-ethereum-flashbots = "==1.65.0"
open-aea-cli-ipfs = "==1.65.0"
clea = "==0.1.0rc4"
cytoolz = "==0.12.3"
docker = "6.1.2"
fastapi = "0.110.3"
eth-hash = "==0.7.0"
eth-account = "==0.8.0"
eth-keyfile = "==0.6.1"
eth-keys = "==0.4.0"
eth-rlp = "==0.3.0"
eth-typing = "==3.5.2"
eth-utils = "==2.3.1"
eth-abi = "==5.1.0"
frozenlist = "==1.4.1"
hexbytes = "==0.3.1"
ipfshttpclient = "==0.8.0a2"
jsonschema = "==4.3.3"
multidict = "==6.0.5"
requests-toolbelt = "1.0.0"
starlette = "0.37.2"
uvicorn = "==0.27.0"
web3 = "6.20.4"
psutil = "^5.9.8"
pyinstaller = "^6.8.0"
aiohttp = "3.9.5"
deepdiff = "^8.0.1"
twikit = "==2.2.0"
halo = "==0.0.31"
argon2-cffi = "==23.1.0"

[tool.poetry.group.development.dependencies]
tomte = {version = "0.2.17", extras = ["cli"]}
build = "1.2.2.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
